notification type and message type enum merge conflict fixes

+++ NotificationApi/NotificationApi.IntegrationTests/Database/Commands/CreateEmailNotificationCommandTests.cs
@@ -4,7 +4,6 @@ using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
 using NotificationApi.DAL;
 using NotificationApi.DAL.Commands;
-using NotificationApi.Domain;
 using NotificationApi.Domain.Enums;
 using NUnit.Framework;
 
@@ -25,7 +24,7 @@ namespace NotificationApi.IntegrationTests.Database.Commands
         [Test]
         public async Task Should_save_new_notification()
         {
-            var notificationType = (int)NotificationType.CreateIndividual;
+            var notificationType = NotificationType.CreateIndividual;
             const string email = "test@email.com";
             var participantId = Guid.NewGuid();
             var hearingId = Guid.NewGuid();
